Fix race condition in SubprocessUtil output handling

+++ compiler/src/org.graalvm.compiler.test/src/org/graalvm/compiler/test/SubprocessUtil.java
@@ -415,8 +415,11 @@ public final class SubprocessUtil {
             });
             outputReader.start();
             boolean finishedOnTime = process.waitFor(timeout.getSeconds(), TimeUnit.SECONDS);
-            int exitCode = process.destroyForcibly().waitFor();
-            return new Subprocess(command, env, exitCode, output, !finishedOnTime);
+            if (!finishedOnTime) {
+                process.destroyForcibly().waitFor();
+            }
+            outputReader.join();
+            return new Subprocess(command, env, process.exitValue(), output, !finishedOnTime);
         }
     }
